style: 添加Send_Packet函数的注释说明

+++ RTT/Debug/frame.h
@@ -15,7 +15,16 @@ typedef enum
     CMD_SET_VAL,
 } Frame_Cmd_t;
 
-void Send_Packet(Frame_Cmd_t cmd, uint8_t *payload, uint16_t len);
+/**
+ * @brief 发送数据包
+ *
+ * 数据包格式：[FRAME_HEAD(2B)][CMD(1B)][PAYLOAD_LEN(1B)][PAYLOAD][CHECKSUM(1B)]
+ *
+ * @param cmd 数据包指令位
+ * @param payload 数据包内容
+ * @param len `payload`的长度
+ */
+void Send_Packet(Frame_Cmd_t cmd, uint8_t *payload, uint8_t len);
 
 void Param_Monitor_Upload();
 void Mapping_Table_Response_Upload();
